Move line counting into XtermTerminal
